Implement Unit Tests and Remove Charging State
We can determine the charging state based on the remaining time of the
battery. When the battery is charging, the state should be None, and
when the battery is discharging, it should have a remaining time.

I have now implemented the mock structure that enables us to isolate the
functions we cannot control. However, we currently need to manually
modify the mock structure whenever we make changes to the
SystemInformation structure. I will explore the use of the 'mockall'
crate to simplify the unit testing process.