Refactor SystemInformation and Use mockall
- Removed the `MockSystemInformation` structure.
- Moved the logic for retrieving system information into a separate
trait called `GetSystemInformation`. Added a default implementation of
the `GetSystemInformation` trait using `DefaultSystemInfoGetter` as the
concrete implementation. Then, updated the `new` function of the
`SystemInformation` structure to utilize the `DefaultSystemInfoGetter`
trait for retrieving disk information and battery remaining time.
- Implemented unit tests using the `mockall` crate. We can mock the
return values of functions in the `GetSystemInformation` trait and pass
the mocked trait into `new_with_sys_info_getter` to test the implemented
logic.

These changes enhance the flexibility and testability of the system
information module.